requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
django-cors-headers==4.3.0
Pillow==10.1.0
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.0
django-phonenumber-field==7.3.0
phonenumbers==8.13.27
gunicorn==21.2.0
whitenoise==6.6.0 